specify locale for Python in test-kitchen
Weirdly never had a problem with this before, until just now...

---

Traceback (most recent call last):
 File "/usr/lib/python3.6/runpy.py", line 193, in _run_module_as_main
   "__main__", mod_spec)
 File "/usr/lib/python3.6/runpy.py", line 85, in _run_code
   exec(code, run_globals)
 File "provisioner.pyz/__main__.py", line 5, in <module>
 File "provisioner.pyz/click/core.py", line 1137, in __call__
 File "provisioner.pyz/click/core.py", line 1043, in main
 File "provisioner.pyz/click/_unicodefun.py", line 100, in _verify_python_env
RuntimeError: Click will abort further execution because Python was configured to use ASCII as encoding for the environment. Consult https://click.palletsprojects.com/unicode-support/ for mitigation steps.

This system supports the C.UTF-8 locale which is recommended. You might be able to resolve your issue by exporting the following environment variables:

   export LC_ALL=C.UTF-8
   export LANG=C.UTF-8

Click discovered that you exported a UTF-8 locale but the locale system could not pick up from it because it does not exist. The exported locale is 'C.UTF-8' but it is not supported.